Create cleaner animation for observing rendering times

+++ app/src/main/java/edu/grinnell/glimmer/nguyenti/mistdroid/Pixel.java
@@ -16,7 +16,7 @@ public class Pixel {
         gradient = Math.random();
     }
 
-    public Pixel(int greyscale) {
+    public Pixel(double greyscale) {
         this.gradient = greyscale;
     }
